requirements.txt
streamlit
ultralytics
opencv-python-headless==4.9.0.80
torch
torchvision
numpy>=1.26.1
pillow
